requirements.txt
flask==3.0.0
garminconnect==0.2.19
python-dotenv==1.0.0
